Implement KUDOS reverse message flow and JUnit test

+++ cf-oscore/src/main/java/org/eclipse/californium/oscore/OscoreOptionDecoder.java
@@ -375,8 +375,8 @@ public class OscoreOptionDecoder {
 	 * 
 	 * @return the y value
 	 */
-	public int getY() {
-		return y;
+	public byte getY() {
+		return (byte) y;
 	}
 
 	/**
